Add support for placeholders in styles (e.g., for font size)
This allows using style definitions with placeholders which can then
be assigned a replacement in the options.styles dict.  Note that TikZ'
append-to-style feature can not cover all cases, in particular, it
cannot be used to (only) change a certain part of a node's font style.
In addition, TikZ append would require the user to "update" all
affected styles.

For now, the default style uses placeholders for the default font size
and font family.