Fix "note" positioning on script actions
The positioning of notes was previously just inherited from regular actions, i.e., the notes were placed left/right of the action node, ignoring the arrow that is present in script actions.

With this commit, a note_left is always placed to the left of the leftmost element of the script action (i.e., the action node if the src is left of the dest, or vice-versa otherwise); and a right_node always to the right of the rightmost element.